refactor(FON-468): migrate admin to feature-first architecture
