+++ apps/client/src/features/administration/components/AdminUserRole.tsx
@@ -2,7 +2,7 @@ import { FormattedMessage } from 'react-intl';
 
 import type { GenderEnum } from '@/types/enums.types';
 
-import { type AdminUserRoleEnum, AdminUserRoleLabel } from './admin-user-enum';
+import { type AdminUserRoleEnum, AdminUserRoleLabel } from '../labels/admin-user-enum';
 
 export const AdminUserRole = (props: { value: AdminUserRoleEnum; gender?: GenderEnum }) => (
   <FormattedMessage {...AdminUserRoleLabel[props.value]} values={{ gender: props.gender }} />
