+++ apps/client/src/features/jobs/types/job.types.ts
