+++ apps/client/src/features/jobs/utils/format-job-duration.utils.ts
