+++ apps/client/src/pages/spaces/admin/jobs-page-outlet-context.type.ts
